requirements.txt
flask == 2.3.3
black == 23.7.0
alembic == 1.12.0
SQLAlchemy == 2.0.20
psycopg2-binary == 2.9.7
python-dotenv == 1.0.0
Flask-Cors == 4.0.0
Flask-SQLAlchemy == 3.1.1